- Vérification d'accès aux fonctions d'ajout / suppression : . Vérification de l'existence des véhicules et les disponibilités concernés . Seulement l'administrateur et le propriétaire peuvent gérer les disponibilités
- Affichage du nom du propriétaire dans le formulaire de disponibilités

+++ www/application/controllers/disponibilite.php
@@ -1,13 +1,15 @@
 <?php
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Contrôleur d'accès aux disponibilités du véhicule
  */
 
 class Disponibilite extends CI_Controller {
 
-    public function create($vehicule_id) {
+    /**
+     * Ajout d'une disponibilité d'un véhicule
+     * @param int $vehicule_id Le véhicule auquel on ajoute la disponibilité
+     */
+    public function create($vehicule_id = NULL) {
 
         // Check login
         if (!UserAcces::userIsLogged()) {
@@ -17,6 +19,7 @@ class Disponibilite extends CI_Controller {
             redirect('noperm');
         }
 
+        // Vérifie l'existence du véhicule
         $vehicule = $this->vehicule_model->getVehiculeById($vehicule_id);
         if (!$vehicule) {
             show_404();
@@ -32,13 +35,20 @@ class Disponibilite extends CI_Controller {
         $data['title'] = 'Disponibilité du véhicule';
         /*
         <?=$base_url?>disponibilite/delete/<?=$disponibilite->getId()?>
-         * 
+         *
          */
         $data['scripts'] = [
             base_url() . 'assets/js/calendrier_date_debut_et_fin.js'
         ];
 
         if ($this->form_validation->run()) {
+
+            // Seulement le propriétaire du véhicule ou l'administrateur
+            //  peut ajouter une disponibilité
+            if (!UserAcces::userIsAdmin() && !UserAcces::getUserId() == $vehicule->getProprietaireId()) {
+                redirect('noperm');
+            }
+
             $this->load->model('disponibilite_model');
             $disponibilite = new EDisponibilite([
                 'vehicule' => $vehicule,
@@ -49,25 +59,39 @@ class Disponibilite extends CI_Controller {
                 $this->session->set_flashdata('msg_success', 'La disponibilité a bien été ajoutée');
                 redirect('disponibilite/create/' . $vehicule_id . '#s');
             }
+            $this->session->set_flashdata('msg_error', 'Un erreur s\'est produite. La disponibilité n\'a pas été ajouté');
         }
 
         $this->load->view('vehicules/disponibilite', $data);
     }
 
+    /**
+     * Supprime une disponibilité
+     * @param int $dispo_id L'identifiant de la disponibilité à supprimer
+     */
     public function delete($dispo_id) {
         $this->load->model('disponibilite_model');
-        $dispo = $this->disponibilite_model->getDisponibiliteById($dispo_id);
 
-        if (!$dispo) {
+        // Vérifie si la disponibilité existe
+        //  et si le véhicule associé existe aussi
+        $dispo = $this->disponibilite_model->getDisponibiliteById($dispo_id);
+        if (!$dispo || !$dispo->getVehicule()) {
             show_404();
         }
 
-        $vehicule_id = $dispo->getVehiculeId();
+        // Seulement le propriétaire du véhicule ou l'administrateur
+        //  peut supprimer une disponibilité
+        $vehicule = $dispo->getVehicule();
+        if (!UserAcces::userIsAdmin() && !UserAcces::getUserId() == $vehicule->getProprietaireId()) {
+            redirect('noperm');
+        }
+
+//        $vehicule_id = $dispo->getVehiculeId();
         if ($this->disponibilite_model->deleteDisponibilite($dispo)) {
             $this->session->set_flashdata('msg_success', 'La disponibilité a été supprimée');
         } else {
             $this->session->set_flashdata('msg_error', 'Un erreur s\'est produite. La disponibilité n\'a pas été supprimée');
         }
-        redirect('disponibilite/create/' . $vehicule_id . '#s');
+        redirect('disponibilite/create/' . $vehicule->getId() . '#s');
     }
 }
